Moved enkf_fs invalidate cache test that need Statoil Data from enkf_fs.c  into separate test enkf_fs_invalidate_cache.c. Rolled back enkf_fs.c to previous version before test for invalidate cache were added. Changed the success critera for enkf_fs_invalidate_cache.c from only non-equal random values to more equal than non-equal values
